HIVE-20121: Investigate issues with TestReplicationScenariosAcidTables (Mahesh Kumar Behera via Sankar Hariappan)
